+++ CHANGELOG.md
@@ -31,7 +31,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     * from `pytorch_lightning.metrics.functional.reduction` removed `reduce`, `class_reduce`	
 
 
-- Removed deprecated `ModelCheckpoint` arguments `prefix`
+- Removed deprecated `ModelCheckpoint` arguments `prefix`, `mode="auto"`
 
 
 ### Fixed